backend/requirements.txt
flask
python-dotenv
flask-restful
flask-sqlalchemy
flask-jwt-extended
flask-mail
flask-cors
flask-login
